view login complete
