+++ client/src/App.css
@@ -1,146 +1,13 @@
-@import url('https://fonts.googleapis.com/css2?family=Nunito+Sans:ital,opsz,wght@0,6..12,200..1000;1,6..12,200..1000&display=swap');
+@import url("https://fonts.googleapis.com/css2?family=Nunito+Sans:ital,opsz,wght@0,6..12,200..1000;1,6..12,200..1000&display=swap");
+@import url("./components/main/notes.css");
+@import url("./components/login/login.css");
+
 
 html,
 body {
   height: 100%;
   margin: 0;
   /* background-color: #eeeeee; */
-  overflow-y: hidden;
-  .notes {
-    display: flex;
-    height: 100%;
-    font-family: sans-serif;
+  overflow-y: hidden;  
 
-    .brand {
-      font-size: 2rem;
-      font-style: oblique;
-      color: antiquewhite;
-      font-weight: 700;
-    }
-    .notes_sidebar {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-      border-right: 1px solid grey;
-      flex-shrink: 0;
-      padding: 1em;
-      max-width: 300px;
-      background-color: orange;
-      overflow-y: auto;
-      height: 100vh;
-      scrollbar-color: rgb(255, 255, 255) rgba(85, 91, 209, 0.2);
-
-      .notes_icon {
-        display: flex;
-        flex-direction: row;
-        align-items: center;
-        font-family: "Nunito Sans", sans-serif;
-        font-size: 1.5rem;
-        font-weight: 800;
-        font-style: italic;
-        letter-spacing: -1px;
-        gap: 2px;
-
-        img {
-          width: 40px;
-          height: 40px;
-        }
-      }
-      .notes_add {
-        background: #009578;
-        border: none;
-        border-radius: 6px;
-        color: #ffff;
-        cursor: pointer;
-        font-size: 1.25em;
-        font-weight: 600;
-        padding: 20px;
-        margin-bottom: 1em;
-        width: 100%;
-        &:hover {
-          background: #00af8c;
-        }
-      }
-
-      .notes_list {
-        display: flex;
-        flex-direction: column;
-        gap: 20px;
-        .notes_list_item {
-          background-color: #eeeeee;
-          border-radius: 7px;
-
-          cursor: pointer;
-          padding: 20px;
-          display: flex;
-          flex-direction: column;
-          align-items: flex-start;
-          gap: 8px;
-
-          &:hover {
-            background-color: #dfdcdc;
-          }
-
-          .notes_small_title {
-            font-size: 1.3rem;
-            font-weight: bold;
-            margin-bottom: 8px;
-          }
-          .notes_small_updated {
-            font-style: italic;
-          }
-        }
-      }
-    }
-    .notes_preview {
-      padding: 2em 3em;
-      flex-grow: 1;
-      margin: 0;
-
-      form {
-        display: flex;
-        flex-direction: column;
-        gap: 30px;
-        align-items: center;
-      }
-      .notes_title,
-      .notes_body {
-        border: none;
-        outline: none;
-        width: 100%;
-        border-radius: 5px;
-        padding: 10px;
-      }
-      .notes_title {
-        font-size: 3em;
-        font-weight: 600;
-        &:focus {
-          text-decoration: underline;
-        }
-      }
-      .notes_body {
-        flex-grow: 1;
-        resize: none;
-        line-height: 1.5;
-        font-size: 1.2em;
-        font-family: sans-serif;
-        height: 55vh;
-      }
-
-      .notes_save {
-        color: white;
-        background-color: rgb(65, 65, 65);
-        max-width: 600px;
-        width: 100%;
-        border: none;
-        border-radius: 6px;
-        font-size: 1.25em;
-        padding: 10px;
-        cursor: pointer;
-        &:hover {
-          background-color: rgb(136, 135, 135);
-        }
-      }
-    }
-  }
 }
