+++ client/src/components/login/login.css
@@ -1,0 +1,69 @@
+.login {
+  width: 100vw;
+  height: 100vh;
+  background-color: orange;
+  place-content: center;
+  font-family: "Nunito Sans", sans-serif;
+
+  .item {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+    background-color: rgb(255, 197, 88);
+    padding: 30px 0;
+
+    .icons {
+      display: flex;
+      gap: 6px;
+      img {
+        width: 50px;
+        height: 50px;
+      }
+      span {
+        font-weight: 700;
+        font-style: italic;
+        font-size: 2rem;
+        text-decoration: underline;
+        letter-spacing: -.5px;
+      }
+    }
+    form {
+      display: flex;
+      flex-direction: column;
+      gap: 30px;
+      align-items: center;
+
+      input {
+        padding: 20px 60px;
+        border: none;
+        outline: none;
+        border-radius: 4px;
+        background-color: rgba(17, 25, 40, 0.6);
+        color: white;
+      }
+      input::placeholder{
+        color: rgb(180, 180, 180);
+      }
+      input:focus{
+        background-color: white;
+        color: black;
+        
+      }
+      button {
+        width: 70%;
+        padding: 10px;
+        outline: none;
+        border: none;
+        border-radius: 4px;
+        background-color: rgb(81, 81, 195);
+        color: white;
+        &:hover {
+            background-color: rgb(67, 67, 163);
+            cursor: pointer;
+        }
+      }
+    }
+  }
+}
